pyproject.toml
[build-system]
requires = [
    "setuptools",
    "setuptools-scm"
]
build-backend = "setuptools.build_meta"

[project]
name = "democorp_airflow"
authors = [{ name = "John Doe", email = "johndoe@democorp.com" }]
description = "Project to demonstrate how to structure an Airflow library with common components."
requires-python = ">=3.9"
classifiers = [
    "Framework :: Apache Airflow",
    "Programming Language :: Python :: 3",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    ]
dependencies = [
    "apache-airflow >= 2.7.0",
    'importlib-metadata; python_version < "3.8"',
    'typing_extensions; python_version < "3.8"',
]
dynamic = ["version", "readme"]

[project.urls]
Homepage = "https://github.com/astronomer/custom-package-demo"

[project.optional-dependencies]
dev = ["black", "isort", "ruff"]
test = ["pytest", "pytest-cov"]
docs = ["sphinx"]

[tool.setuptools.dynamic]
readme = { file = ["README.md"] }

[tool.ruff]
# https://docs.astral.sh/ruff/configuration/#using-pyprojecttoml
line-length = 110

[tool.ruff.lint]
select = [
    "E", # pycodestyle
    "F", # Pyflakes
    "UP", # pyupgrade
    "B", # flake8-bugbear
    "SIM", # flake8-simplify
    "I", # isort
]

[tool.setuptools.packages]
find = {}

[tool.setuptools_scm]
